+++ torch_similarity_search/__init__.py
@@ -3,6 +3,7 @@
 from torch_similarity_search.converters.faiss_converter import from_faiss
 from torch_similarity_search.indexes.flat import FlatIndex
 from torch_similarity_search.indexes.ivf_flat import IVFFlatIndex
+from torch_similarity_search.indexes.ivf_pq import IVFPQIndex
 
-__all__ = ["FlatIndex", "IVFFlatIndex", "from_faiss"]
-__version__ = "0.0.2"
+__all__ = ["FlatIndex", "IVFFlatIndex", "IVFPQIndex", "from_faiss"]
+__version__ = "0.0.3"